+++ admin-app/src/feature/schedule/ScheduleForm.tsx
@@ -1,7 +1,7 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Button } from "@yamada-ui/react";
 import { useForm, FormProvider } from "react-hook-form";
-import { InputField } from "../../components/InputField";
+import { InputField } from "../../components/form/InputField";
 import { scheduleSchema } from "./regisrter/scheduleSchema";
 import { createSchedule } from "./api";
