+++ admin-app/src/routes/Movies.tsx
@@ -1,4 +1,4 @@
-import { AddMovieModal } from "../components/AddMovieModal";
+import { AddMovieModal } from "../components/modal/AddMovieModal";
 import { MovieTable } from "../data-table/MovieTable";
 
 export const Movies = () => {
