+++ admin-app/src/routes/Users.tsx
@@ -1,5 +1,5 @@
-import { AddUserModal } from "../components/AddUserModal";
-import { UserTable } from "../data-table/UserTable"
+import { AddUserModal } from "../components/modal/AddUserModal";
+import { UserTable } from "../data-table/UserTable";
 
 export const Users = () => {
   return (
@@ -7,5 +7,5 @@ export const Users = () => {
       <AddUserModal />
       <UserTable />
     </>
-  )
-}
+  );
+};
