+++ mock/hooks/useUsers.tsx
@@ -1,14 +1,14 @@
 import { useEffect, useState } from "react";
 import { UserMock } from "../types/user";
 import { apiGet } from "../../util/apiClient";
-export const useUser = (pageId?: string, pageSize?: string) => {
-  const [user, setUser] = useState<UserMock[]>([]);
+export const useUsers = (pageId?: string, pageSize?: string) => {
+  const [users, setUsers] = useState<UserMock[]>([]);
 
   const fetchData = async (pageId?: string, pageSize?: string) => {
     try {
       const url = "http://localhost:8014/users";
       const res = await apiGet(url + `?pageId=${pageId}&pageSize=${pageSize}`);
-      setUser(res);
+      setUsers(res);
     } catch (error) {
       console.error("user service error: " + error);
     }
@@ -18,5 +18,5 @@ export const useUser = (pageId?: string, pageSize?: string) => {
     fetchData(pageId, pageSize);
   }, [pageId, pageSize]);
 
-  return { user, setUser };
+  return { users, setUsers };
 };
